refactor(actions): move actions to separated file
